requirements.txt
bottle
python-mpd2
pillow
gevent-websocket
